backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
bcrypt==4.0.1
    # via
    #   event-link-backend (pyproject.toml)
    #   passlib
certifi==2024.8.30
    # via httpx
alembic==1.14.0
    # via event-link-backend (pyproject.toml)
cffi==2.0.0
    # via cryptography
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via python-jose
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via pydantic
fastapi==0.121.2
    # via event-link-backend (pyproject.toml)
greenlet==3.2.4
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.28.1
    # via event-link-backend (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   httpx
    #   email-validator
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
passlib==1.7.4
    # via event-link-backend (pyproject.toml)
psycopg2-binary==2.9.11
    # via event-link-backend (pyproject.toml)
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   event-link-backend (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via event-link-backend (pyproject.toml)
python-dotenv==1.2.1
    # via pydantic-settings
python-jose==3.5.0
    # via event-link-backend (pyproject.toml)
python-multipart==0.0.20
    # via event-link-backend (pyproject.toml)
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via event-link-backend (pyproject.toml)
starlette==0.49.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.38.0
    # via event-link-backend (pyproject.toml)



pytest==8.3.3
pytest-cov==5.0.0
